Doxygen: Do not include .rst files but do include .txt files
The .rst files were ignored anyway since rst is not
natively supported and there was no mapping defined.

But include .txt files and parse them as MarkDown.
The handles e.g. the files under LICENSES/ and works
reasonably well.

Need to extend the exclude patterns though to exclude
UT inputs which are generally not useful and sometimes
nasty (on purpose).

Signed-off-by: Frank Lichtenheld <frank@lichtenheld.com>